:tada: The shared folder can now be selected
Before thumbcloud would just share the folder it was started in. Now the
user can select which folder will be shared. I am really excited about
this commit because now thumbcloud is somehow useable if you ignore the
fact that files cannot be uploaded yet. Moreover, we are now near the
0.0.2 release.