Fix relative paths not working
As mentioned in the #4 issue relative paths were not working.

This was because Thumbcloud checks at every request from the client if
the client tries to access files it shouldn't. It does this by simple
comparison. However it got very confused if a relative path got entered
because it didn't converted the relative path to an absolute path.

Now it does this convertion and relative paths work just fine.